requirements.txt
fastapi
openai
python-dotenv
python-multipart
pymongo
aiofiles
sseclient